Create README.md
In this case study, we are attempting to solve a real-world business problem using Exploratory Data Science techniques. 
We will be understanding and solving a risk analytics problem in Banking and Financial Domain.
We will be checking how data can be used effectively to solve business problems like defaulter’s prediction in Loan Lending club.